DocumentWriter API simplified to proxy role. PDF writing logic moved to PdfDocumentWriter.
